Lint fixes from GitHub CI
