+++ internal/service/resiliencehubv2/assertion_test.go
@@ -84,6 +84,8 @@ func testAccCheckAssertionExists(ctx context.Context, t *testing.T, n string, v 
 
 func testAccAssertionConfig_basic(rName string) string {
 	return fmt.Sprintf(`
+data "aws_region" "current" {}
+
 resource "aws_resiliencehubv2_policy" "test" {
   name = "%[1]s-policy"
 
@@ -94,7 +96,7 @@ resource "aws_resiliencehubv2_policy" "test" {
 
 resource "aws_resiliencehubv2_service" "test" {
   name    = "%[1]s-service"
-  regions = ["us-west-2"]
+  regions = [data.aws_region.current.name]
 
   policy_arn = aws_resiliencehubv2_policy.test.arn
