+++ internal/service/resiliencehubv2/complex_test.go
@@ -60,6 +60,8 @@ func TestAccResilienceHubV2_complex(t *testing.T) {
 
 func testAccComplexConfig(rName string) string {
 	return fmt.Sprintf(`
+data "aws_region" "current" {}
+
 resource "aws_resiliencehubv2_policy" "test" {
   name = "%[1]s-policy"
 
@@ -81,7 +83,7 @@ resource "aws_resiliencehubv2_system" "test" {
 
 resource "aws_resiliencehubv2_service" "test" {
   name    = "%[1]s-service"
-  regions = ["us-west-2"]
+  regions = [data.aws_region.current.name]
 
   policy_arn = aws_resiliencehubv2_policy.test.arn
