+++ internal/service/resiliencehubv2/policy_data_source_test.go
@@ -1,0 +1,55 @@
+// Copyright IBM Corp. 2014, 2026
+// SPDX-License-Identifier: MPL-2.0
+
+package resiliencehubv2_test
+
+import (
+	"fmt"
+	"testing"
+
+	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
+	"github.com/hashicorp/terraform-provider-aws/internal/acctest"
+	"github.com/hashicorp/terraform-provider-aws/names"
+)
+
+func TestAccResilienceHubV2PolicyDataSource_basic(t *testing.T) {
+	ctx := acctest.Context(t)
+	rName := acctest.RandomWithPrefix(t, "tf-acc-test")
+	resourceName := "aws_resiliencehubv2_policy.test"
+	dataSourceName := "data.aws_resiliencehubv2_policy.test"
+
+	acctest.ParallelTest(ctx, t, resource.TestCase{
+		PreCheck: func() {
+			acctest.PreCheck(ctx, t)
+			testAccPreCheck(ctx, t)
+		},
+		ErrorCheck:               acctest.ErrorCheck(t, names.ResilienceHubV2),
+		ProtoV5ProviderFactories: acctest.ProtoV5ProviderFactories,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccPolicyDataSourceConfig_basic(rName),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttrPair(dataSourceName, names.AttrARN, resourceName, names.AttrARN),
+					resource.TestCheckResourceAttrPair(dataSourceName, names.AttrName, resourceName, names.AttrName),
+					resource.TestCheckResourceAttr(dataSourceName, "availability_slo.0.target", "99.9"),
+				),
+			},
+		},
+	})
+}
+
+func testAccPolicyDataSourceConfig_basic(rName string) string {
+	return fmt.Sprintf(`
+resource "aws_resiliencehubv2_policy" "test" {
+  name = %[1]q
+
+  availability_slo {
+    target = 99.9
+  }
+}
+
+data "aws_resiliencehubv2_policy" "test" {
+  arn = aws_resiliencehubv2_policy.test.arn
+}
+`, rName)
+}
