+++ internal/service/resiliencehubv2/service_function_test.go
@@ -85,6 +85,8 @@ func testAccCheckServiceFunctionExists(ctx context.Context, t *testing.T, n stri
 
 func testAccServiceFunctionConfig_basic(rName string) string {
 	return fmt.Sprintf(`
+data "aws_region" "current" {}
+
 resource "aws_resiliencehubv2_policy" "test" {
   name = "%[1]s-policy"
 
@@ -95,7 +97,7 @@ resource "aws_resiliencehubv2_policy" "test" {
 
 resource "aws_resiliencehubv2_service" "test" {
   name    = "%[1]s-service"
-  regions = ["us-west-2"]
+  regions = [data.aws_region.current.name]
 
   policy_arn = aws_resiliencehubv2_policy.test.arn
