+++ internal/service/resiliencehubv2/service_test.go
@@ -96,6 +96,8 @@ func testAccCheckServiceExists(ctx context.Context, t *testing.T, n string, v *a
 
 func testAccServiceConfig_basic(rName string) string {
 	return fmt.Sprintf(`
+data "aws_region" "current" {}
+
 resource "aws_resiliencehubv2_policy" "test" {
   name = "%[1]s-policy"
 
@@ -106,7 +108,7 @@ resource "aws_resiliencehubv2_policy" "test" {
 
 resource "aws_resiliencehubv2_service" "test" {
   name    = %[1]q
-  regions = ["us-west-2"]
+  regions = [data.aws_region.current.name]
 
   policy_arn = aws_resiliencehubv2_policy.test.arn
