+++ internal/service/resiliencehubv2/testdata/Service/region_override/main_gen.tf
@@ -10,6 +10,8 @@ resource "aws_resiliencehubv2_policy" "test" {
 }
 
 resource "aws_resiliencehubv2_service" "test" {
+  region = var.region
+
   name    = var.rName
   regions = ["us-west-2"]
