+++ internal/service/resiliencehubv2/testdata/tmpl/policy_basic.gtpl
@@ -1,4 +1,5 @@
 resource "aws_resiliencehubv2_policy" "test" {
+{{- template "region" }}
   name = var.rName
 
   availability_slo {
