+++ internal/service/resiliencehubv2/testdata/tmpl/system_basic.gtpl
@@ -1,4 +1,5 @@
 resource "aws_resiliencehubv2_system" "test" {
+{{- template "region" }}
   name = var.rName
 
 {{- template "tags" . }}
