+++ website/docs/d/resiliencehubv2_system.html.markdown
@@ -25,6 +25,7 @@ data "aws_resiliencehubv2_system" "example" {
 This data source supports the following arguments:
 
 * `arn` - (Required) ARN of the system.
+* `region` - (Optional, **Deprecated**) Region where this resource will be [managed](https://docs.aws.amazon.com/general/latest/gr/rande.html#regional-endpoints). Defaults to the Region set in the [provider configuration](https://registry.terraform.io/providers/hashicorp/aws/latest/docs#aws-configuration-reference).
 
 ## Attribute Reference
