+++ website/docs/r/resiliencehubv2_assertion.html.markdown
@@ -30,6 +30,10 @@ The following arguments are required:
 * `service_arn` - (Required) ARN of the service this assertion belongs to. Changing this value requires creating a new resource.
 * `text` - (Required) Text of the resilience assertion.
 
+The following arguments are optional:
+
+* `region` - (Optional, **Deprecated**) Region where this resource will be [managed](https://docs.aws.amazon.com/general/latest/gr/rande.html#regional-endpoints). Defaults to the Region set in the [provider configuration](https://registry.terraform.io/providers/hashicorp/aws/latest/docs#aws-configuration-reference).
+
 ## Attribute Reference
 
 This resource exports the following attributes in addition to the arguments above:
