+++ website/docs/r/resiliencehubv2_service_function.html.markdown
@@ -46,6 +46,7 @@ The following arguments are required:
 The following arguments are optional:
 
 * `description` - (Optional) Description of the service function.
+* `region` - (Optional, **Deprecated**) Region where this resource will be [managed](https://docs.aws.amazon.com/general/latest/gr/rande.html#regional-endpoints). Defaults to the Region set in the [provider configuration](https://registry.terraform.io/providers/hashicorp/aws/latest/docs#aws-configuration-reference).
 
 ## Attribute Reference
