+++ website/docs/r/resiliencehubv2_user_journey.html.markdown
@@ -61,6 +61,7 @@ The following arguments are optional:
 
 * `description` - (Optional) Description of the user journey.
 * `policy_arn` - (Optional) ARN of the resilience policy to associate with this user journey.
+* `region` - (Optional, **Deprecated**) Region where this resource will be [managed](https://docs.aws.amazon.com/general/latest/gr/rande.html#regional-endpoints). Defaults to the Region set in the [provider configuration](https://registry.terraform.io/providers/hashicorp/aws/latest/docs#aws-configuration-reference).
 
 ## Attribute Reference
